+++ apps/react-demo/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @formulaxjs/react-demo
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @formulaxjs/tiptap@0.5.0
+
 ## 0.1.2
 
 ### Patch Changes