+++ apps/vue-demo/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @formulaxjs/vue-demo
 
+## 0.1.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @formulaxjs/tinymce@0.4.0
+
 ## 0.1.2
 
 ### Patch Changes